Fix brightness type handling and enable strict typing for light.py

+++ homeassistant/components/lutron/const.py
@@ -4,4 +4,4 @@ DOMAIN = "lutron"
 
 CONF_DEFAULT_DIMMER_LEVEL = "default_dimmer_level"
 
-DEFAULT_DIMMER_LEVEL = 255 / 2
+DEFAULT_DIMMER_LEVEL = int(255 / 2)
